[refactor] rename mmap_parse_(16|32) to decode_inst_(16|32)
